# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for the HS project",
        "title": "HS Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/prisms/{prismId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a prism",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prisms"
                ],
                "summary": "Delete prism",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Prism ID",
                        "name": "prismId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a prism",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prisms"
                ],
                "summary": "Update prism",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Prism ID",
                        "name": "prismId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Prism",
                        "name": "prism",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePrismInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePrismResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/users/communication-services": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns user communication services",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Communication Services"
                ],
                "summary": "Get user communication services",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.GetUserCommunicationServicesResponseItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new user communication service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Communication Services"
                ],
                "summary": "Create user communication service",
                "parameters": [
                    {
                        "description": "User Communication Service",
                        "name": "userCommunicationService",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserCommunicationServiceInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserCommunicationServiceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/users/communication-services/{userCommunicationServiceId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a user communication service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Communication Services"
                ],
                "summary": "Delete user communication service",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User Communication Service ID",
                        "name": "userCommunicationServiceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a user communication service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Communication Services"
                ],
                "summary": "Update user communication service",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User Communication Service ID",
                        "name": "userCommunicationServiceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User Communication Service",
                        "name": "userCommunicationService",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserCommunicationServiceInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserCommunicationServiceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateUserCommunicationServiceInput": {
            "type": "object",
            "required": [
                "name",
                "service",
                "value"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/enum.CommunicationService"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "dto.CreateUserCommunicationServiceResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/enum.CommunicationService"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetUserCommunicationServicesResponseItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/enum.CommunicationService"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdatePrismEnrichedConfig": {
            "type": "object",
            "properties": {
                "base": {
                    "$ref": "#/definitions/dto.UpdatePrismFacet"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UpdatePrismEnrichedUserItem"
                    }
                }
            }
        },
        "dto.UpdatePrismEnrichedUserItem": {
            "type": "object",
            "properties": {
                "facet": {
                    "$ref": "#/definitions/dto.UpdatePrismFacet"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "dto.UpdatePrismFacet": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "privateLabel": {
                    "type": "string"
                },
                "publicLabel": {
                    "type": "string"
                }
            }
        },
        "dto.UpdatePrismInput": {
            "type": "object",
            "properties": {
                "configuration": {
                    "$ref": "#/definitions/json.PrismConfig"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.UpdatePrismResponse": {
            "type": "object",
            "properties": {
                "configuration": {
                    "$ref": "#/definitions/dto.UpdatePrismEnrichedConfig"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateUserCommunicationServiceInput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/enum.CommunicationService"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateUserCommunicationServiceResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/enum.CommunicationService"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "enum.CommunicationService": {
            "type": "string",
            "enum": [
                "faceToFace",
                "phone",
                "message",
                "email",
                "discord",
                "microsoftTeams"
            ],
            "x-enum-varnames": [
                "FaceToFace",
                "Phone",
                "Message",
                "Email",
                "Discord",
                "MicrosoftTeams"
            ]
        },
...
        "error.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "json.PrismConfig": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/json.PrismConfigUserItem"
                    }
                }
            }
        },
        "json.PrismConfigUserItem": {
            "type": "object",
            "properties": {
                "facetId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}